refactor: replace legacy toast implementation with Sonner library
- Updated toast notifications across components to use the Sonner library for improved functionality and consistency.
- Removed the old useToast hook and related components.
- Adjusted toast usage in various components to align with the new API, ensuring error and success messages are displayed correctly.
- Enhanced the Toaster component with new configuration options for better user experience.
