+++ client/src/components/AccountMenu/index.tsx
@@ -15,22 +15,17 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useAuth } from "@/contexts";
-import { useToast } from "@/hooks/use-toast";
+import { toast } from "@/lib/toast";
 
 export function AccountMenu(): ReactElement {
   const { user, logout } = useAuth();
-  const { toast } = useToast();
   const [openEditUserDialog, setOpenEditUserDialog] = useState(false);
   const [logoutMutation] = useLogoutMutation({
     onCompleted() {
       logout?.();
     },
     onError: (error) => {
-      toast({
-        title: "Error",
-        description: `Logout: ${error.message}`,
-        variant: "destructive",
-      });
+      toast.error(`Logout: ${error.message}`);
     },
   });
