+++ client/src/components/Deck/index.tsx
@@ -4,7 +4,7 @@ import { usePickCardMutation } from "@/api";
 import { Card } from "@/components/Card";
 import { useAuth } from "@/contexts";
 import { useKeyboardControls } from "@/hooks";
-import { useToast } from "@/hooks/use-toast";
+import { toast } from "@/lib/toast";
 import { cn } from "@/lib/utils";
 import { UserCard } from "@/types";
 import { getPickedUserCard } from "@/utils";
@@ -24,17 +24,12 @@ export function Deck({
 }: DeckProps): ReactElement {
   const [selectedCard, setSelectedCard] = useState<string | null>(null);
   const { user } = useAuth();
-  const { toast } = useToast();
   const { cardsContainerRef } = useKeyboardControls();
 
   const [pickCardMutation] = usePickCardMutation({
     onError(error) {
       setSelectedCard(null);
-      toast({
-        title: "Error",
-        description: `Pick card: ${error.message}`,
-        variant: "destructive",
-      });
+      toast.error(`Pick card: ${error.message}`);
     },
   });
