+++ client/src/hooks/useCopyRoomUrlToClipboard.ts
@@ -1,8 +1,7 @@
 import { useCallback, useMemo } from "react";
 
+import { toast } from "@/lib/toast";
 import { copyTextToClipboard } from "@/utils";
-
-import { toast } from "./use-toast";
 
 interface UseCopyRoomUrlReturn {
   copyRoomUrlToClipboard: (roomId: string) => Promise<void>;
@@ -16,17 +15,11 @@ export function useCopyRoomUrlToClipboard(): UseCopyRoomUrlReturn {
     const isCopySuccess = await copyTextToClipboard(`${origin}${roomPath}`);
 
     if (isCopySuccess) {
-      toast({
-        title: "Invite link copied to clipboard",
-        variant: "default",
-      });
+      toast.success("Invite link copied to clipboard");
     } else {
-      toast({
-        title: "Error",
-        description:
-          "When copying a invite link something went wrong. But don't be discouraged, just copy it yourself from the browser.",
-        variant: "destructive",
-      });
+      toast.error(
+        "When copying a invite link something went wrong. But don't be discouraged, just copy it yourself from the browser.",
+      );
     }
   }, []);
